+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - L'action `Extraction de l'archive` étant automatique, le bouton n'est plus visible pour le flux `Pièces de marchés par étapes` #1835
 - Il est possible de générer des UUID à la place des 7 caractères actuels pour les identifiants des nouveaux dossiers #1098
+- Le script installation/test-last-job.php est remplacé par la commande `bin/console app:daemon:notify-check` #1810
 
 ## Corrections